requirements.txt
protobuf~=5.29.4
slack_bolt~=1.23.0
python-dotenv~=1.1.0
slack_bolt~=1.23.0
slack_sdk~=3.35.0
openai~=1.82.0
slack_bolt~=1.23.0
slack_sdk~=3.35.0
chromadb~=1.0.10
slack_bolt~=1.23.0
slack_sdk~=3.35.0
tiktoken~=0.9.0
langchain-community~=0.3.24
Markdown~=3.8
PyPDF2~=3.0.1
beautifulsoup4~=4.13.4
slack_bolt~=1.23.0
slack_sdk~=3.35.0
langchain-text-splitters~=0.3.8
slack_bolt~=1.23.0
slack_sdk~=3.35.0
langchain-core~=0.3.61
slack_bolt~=1.23.0
slack_sdk~=3.35.0
langchain~=0.3.25
langchain-google-genai~=2.1.4